+++ test/src/flutter_floaty_test.dart
@@ -15,8 +15,8 @@ void main() {
               FlutterFloaty(
                 initialX: 50,
                 initialY: 50,
-                initialWidth: 200,
-                initialHeight: 200,
+                width: 200,
+                height: 200,
                 builder: (context) => const Text('Floaty'),
               ),
             ],
@@ -25,7 +25,13 @@ void main() {
       ),
     );
 
+    // Check if the widget is present
     expect(find.text('Floaty'), findsOneWidget);
+
+    // Check if the widget has the correct initial size
+    final floaty = tester.firstWidget(find.byType(Container)) as Container;
+    expect(floaty.constraints!.maxWidth, 200);
+    expect(floaty.constraints!.maxHeight, 200);
   });
 
   testWidgets('FlutterFloaty responds to tap gestures',
@@ -40,8 +46,8 @@ void main() {
               FlutterFloaty(
                 initialX: 50,
                 initialY: 50,
-                initialWidth: 200,
-                initialHeight: 200,
+                width: 200,
+                height: 200,
                 builder: (context) => const Text('Floaty'),
                 onTap: () {
                   tapped = true;
@@ -67,8 +73,8 @@ void main() {
               FlutterFloaty(
                 initialX: 50,
                 initialY: 50,
-                initialWidth: 200,
-                initialHeight: 200,
+                width: 200,
+                height: 200,
                 builder: (context) => const Text('Floaty'),
               ),
             ],
@@ -85,5 +91,36 @@ void main() {
     expect(newPosition, isNot(initialPosition));
     expect(newPosition.dx, initialPosition.dx + 200);
     expect(newPosition.dy, initialPosition.dy + 200);
+  });
+
+  testWidgets('FlutterFloaty respects intrinsic boundaries',
+      (WidgetTester tester) async {
+    await tester.pumpWidget(
+      MaterialApp(
+        home: Scaffold(
+          body: Stack(
+            children: [
+              FlutterFloaty(
+                initialX: 50,
+                initialY: 50,
+                width: 200,
+                height: 200,
+                intrinsicBoundaries: const Rect.fromLTWH(0, 0, 300, 300),
+                builder: (context) => const Text('Floaty'),
+              ),
+            ],
+          ),
+        ),
+      ),
+    );
+
+    final initialPosition = tester.getTopLeft(find.text('Floaty'));
+    await tester.drag(find.text('Floaty'), const Offset(300, 300));
+    await tester.pumpAndSettle();
+    final newPosition = tester.getTopLeft(find.text('Floaty'));
+
+    // Ensure the widget doesn't go out of the boundaries (300, 300 is the max)
+    expect(newPosition.dx, lessThanOrEqualTo(200)); // 300 - 200 (width)
+    expect(newPosition.dy, lessThanOrEqualTo(200)); // 300 - 200 (height)
   });
 }